% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_chromosome_size}
\alias{add_chromosome_size}
\title{Add chromosome size to a data frame}
\usage{
add_chromosome_size(data, chromSize = NULL, karyotype = "hg19")
}
\arguments{
\item{data}{A data frame with columns Chrom, Start, End}

\item{chromSize}{A data frame with columns Chrom, Size}

\item{karyotype}{A character string specifying the karyotype to use}
}
\value{
A data frame with an additional column size
}
\description{
Add chromosome size to a data frame
}
