% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome.size.hg19}
\alias{chromosome.size.hg19}
\title{chromosome size data for hg19}
\format{
a data.frame with 24 rows and 2 variables:
\describe{
\item{Chrom}{chromosome (chr1-22, chrX, chrY)}
\item{Start}{chromosome start position}
\item{End}{chromosome end position}
}
}
\usage{
chromosome.size.hg19
}
\description{
chromosome size data for hg19
}
\keyword{datasets}
