% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{cytoband_colors}
\alias{cytoband_colors}
\title{Get cytoband colors}
\format{
An object of class \code{character} of length 10.
}
\usage{
cytoband_colors
}
\value{
A named vector with cytoband colors
}
\description{
Get cytoband colors
}
\keyword{datasets}
