% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ideogram.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{Ideogram}
\usage{
geom_ideogram(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  radius = unit(0.1, "npc"),
  chrom.col = NULL,
  chrom.lwd = NULL,
  chrom.lty = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by ggplot2::aes() or ggplot2::aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{radius}{Radius of rounded corners.}

\item{chrom.col}{Color of the chromosome.}

\item{chrom.lwd}{Line width of the chromosome.}

\item{chrom.lty}{Line type of the chromosome.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().}
}
\description{
draw ideogram geom
}
\details{
ideogram with ggplot2
}
\examples{
library(ggplot2)
library(ggideogram)

region <- seq(0.0, 0.5, length.out = 48)
band_data <- data.frame(
  x = rep(1, 48),
  ymin = region,
  ymax = region + diff(region)[1],
  fill = c("cyan", "purple", "cyan", "purple", "cyan", "purple"),
  arm = rep(c("p", "q"), each=24)
)

ggplot(band_data, aes(chrom = 1)) +
  geom_ideogram(aes(x = x, ymin = ymin, ymax = ymax, arm = arm, fill = fill),
                radius = unit(20, 'pt')) +
  theme_void()
}
