% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karyotype_hg19}
\alias{karyotype_hg19}
\title{human genome hg19 karyotype data}
\format{
a data.frame with 862 rows and 5 variables:
\describe{
\item{Chrom}{chromosome (chr1-22, chrX, chrY)}
\item{Start}{start position (0)}
\item{End}{end position}
\item{Stain}{chromosomal bands}
\item{Arm}{long arm or short arm (p or q)}
}
}
\usage{
karyotype_hg19
}
\description{
human genome hg19 karyotype data
}
\keyword{datasets}
