% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{put_togather}
\alias{put_togather}
\title{Put together ideogram data}
\usage{
put_togather(data, regions = NULL, width = 1e+06)
}
\arguments{
\item{data}{A data frame with columns Chrom, Start, End, ...}

\item{regions}{A data frame with columns Chrom, Size. If NULL, it will be set to data.
If not NULL, it will be used to calculate the increment for each chromosome.
The chromosome should be the same as in data and ordered in the same way.}

\item{width}{The width of the ideogram}
}
\value{
A data frame with the ideogram data put together
}
\description{
Put together ideogram data
}
