% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband.R
\name{read_cytoband}
\alias{read_cytoband}
\title{Read a UCSC cytoband table}
\usage{
read_cytoband(file, chr = NULL)
}
\arguments{
\item{file}{A path, URL, or data frame containing the five positional UCSC
columns: chromosome, start, end, band name and stain.}

\item{chr}{Optional chromosome selection and order.}
}
\value{
A data frame with \code{Chr}, \code{Start}, \code{End}, \code{Name}, and \code{Stain}.
}
\description{
Read a UCSC cytoband table
}
