% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-inset.R
\name{as_ideogram_grob}
\alias{as_ideogram_grob}
\title{Convert an ideogram to a standard grob}
\usage{
as_ideogram_grob(plot)
}
\arguments{
\item{plot}{A [ggideogram()] plot.}
}
\value{
A standard gtable/grob.
}
\description{
This is a stable convenience wrapper around [ggplot2::ggplotGrob()] and does
not create a second renderer or fixed page layout.
}
