% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_karyotype}
\alias{human_karyotype}
\title{Human karyotype}
\format{
A data frame with 24 rows and 5 columns:
\describe{
  \item{Chr}{Chromosome name.}
  \item{Start, End}{Chromosome extent, in base pairs.}
  \item{CE_start, CE_end}{Centromere extent, in base pairs.}
}
}
\source{
RIdeogram: \url{https://github.com/TickingClock1992/RIdeogram}
}
\usage{
human_karyotype
}
\description{
The 24 human chromosomes with centromere positions, in the five-column form
accepted by [ggideogram()].
}
\keyword{datasets}
